PC file storage is removed. All required variables, methods and Storage inerface moved to other files.
